requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# tester/geth.py: 4
PyYAML == 3.12

# main.py: 1,2
baelfire == 0.5.1

# tester/statistics_collector.py: 1
pandas == 0.21.0
